+++ src/main/java/insane96mcp/iguanatweaksexpanded/network/message/JumpMidAirMessage.java
@@ -1,6 +1,6 @@
 package insane96mcp.iguanatweaksexpanded.network.message;
 
-import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.MAJump;
+import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.DoubleJump;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraftforge.network.NetworkDirection;
@@ -25,7 +25,7 @@ public class JumpMidAirMessage {
 		ctx.get().enqueueWork(() -> {
 			if (ctx.get().getSender() != null) {
 				ctx.get().getSender().jumpFromGround();
-				MAJump.playAnimation(ctx.get().getSender());
+				DoubleJump.playAnimation(ctx.get().getSender());
 			}
 		});
 		ctx.get().setPacketHandled(true);